Refactor ImportSettingsDialog.tsx to extract utilities
Extracted utility functions (`detectDelimiter`, `detectDecimalPoint`, `detectColumnTypeAndFormat`, `generatePreviewData`, `generateColumnConfigs`, and `getPreferredXAxisColumn`) from `ImportSettingsDialog.tsx` into standalone utility modules under `ImportSettings/utils`. Tests for these utilities were also added, achieving excellent coverage.

This improves maintainability by reducing the size of the React component file and separating core data parsing/detecting logic from UI code.
