+++ src/components/Layout/ImportSettingsDialog.tsx
@@ -5,10 +5,10 @@ import type {
 	ColumnType,
 	ImportSettings,
 } from "../../types/import";
-import { splitCSVLine } from "../../utils/data-parser";
-import { secureJSONParse } from "../../utils/json";
 import { PopupPicker, type PopupPickerOption } from "../Sidebar/PopupPicker";
 import { Modal } from "./Modal";
+import { detectDelimiter, detectDecimalPoint } from "./ImportSettings/utils/detectors";
+import { generatePreviewData, generateColumnConfigs, getPreferredXAxisColumn } from "./ImportSettings/utils/preview";
 
 const TYPE_OPTIONS_META = [
 	{ type: "numeric" as const, icon: Hash, label: "Numeric" },
@@ -33,246 +33,6 @@ interface ImportSettingsDialogProps {
 	onSheetChange?: (sheet: string) => void;
 	onConfirm: (settings: ImportSettings) => void | Promise<void>;
 	onCancel: () => void;
-}
-
-function calculateDelimiterScore(lines: string[], delimiter: string): number {
-	let totalCount = 0;
-	const counts = new Map<number, number>();
-	let maxConsistency = 0;
-
-	for (const line of lines) {
-		const count = line.split(delimiter).length - 1;
-		totalCount += count;
-		if (count > 0) {
-			const c = (counts.get(count) || 0) + 1;
-			counts.set(count, c);
-			maxConsistency = Math.max(maxConsistency, c);
-		}
-	}
-
-	// Score: number of lines with the most consistent non-zero count,
-	// plus a small bonus for total count to break ties.
-	return maxConsistency * 1000 + totalCount;
-}
-
-function detectDelimiter(
-	fileContent: string,
-	fileType: "csv" | "json" | "excel",
-): string {
-	if (fileType !== "csv" && fileType !== "excel") return ",";
-	if (!fileContent) return ",";
-
-	const lines = fileContent
-		.split(/\r?\n/)
-		.slice(0, 100)
-		.filter((l) => l.trim());
-	if (lines.length === 0) return ",";
-
-	const candidates = [",", ";", "\t", "|"];
-	let best = ",";
-	let maxScore = -1;
-
-	for (const d of candidates) {
-		const score = calculateDelimiterScore(lines, d);
-		if (score > maxScore) {
-			maxScore = score;
-			best = d;
-		}
-	}
-	return best;
-}
-
-function detectDecimalPoint(fileContent: string, delimiter: string): string {
-	if (!fileContent) return ".";
-	const lines = fileContent
-		.split(/\r?\n/)
-		.slice(0, 100)
-		.filter((l) => l.trim());
-	let dotCount = 0;
-	let commaCount = 0;
-
-	const dotRegex = /\d\.\d/g;
-	const commaRegex = /\d,\d/g;
-
-	for (const line of lines) {
-		const fields = line.split(delimiter);
-		for (const field of fields) {
-			dotCount += (field.match(dotRegex) || []).length;
-			commaCount += (field.match(commaRegex) || []).length;
-		}
-	}
-
-	return commaCount > dotCount ? "," : ".";
-}
-
-function detectColumnTypeAndFormat(
-	firstVal: string | undefined,
-	decimalPoint: string,
-): { type: ColumnType; dateFormat?: string } {
-	if (!firstVal || firstVal.trim() === "") {
-		return { type: "ignore" };
-	}
-
-	let type: ColumnType = "numeric";
-	let dateFormat: string | undefined;
-
-	const normalized = firstVal.replace(decimalPoint, ".");
-	if (Number.isNaN(Number(normalized)) || normalized.split(".").length > 2) {
-		if (
-			firstVal.includes("-") ||
-			firstVal.includes(".") ||
-			firstVal.includes("/")
-		) {
-			type = "date";
-			if (firstVal.match(/^\d{4}-\d{2}-\d{2}$/)) dateFormat = "YYYY-MM-DD";
-			else if (firstVal.match(/^\d{2}\.\d{2}\.\d{4}$/))
-				dateFormat = "DD.MM.YYYY";
-			else if (firstVal.match(/^\d{2}\/\d{2}\/\d{4}$/))
-				dateFormat = "DD/MM/YYYY";
-		} else {
-			type = "categorical";
-		}
-	}
-	return { type, dateFormat };
-}
-
-interface PreviewData {
-	headers: string[];
-	rows: Record<string, string>[] | string[][];
-	skippedLines: string[];
-	gapStart: number | null;
-	totalRows: number;
-}
-
-function generatePreviewData(
-	fileContent: string,
-	fileType: "csv" | "json" | "excel",
-	delimiter: string,
-	startRow: number,
-	commentChar: string,
-): PreviewData {
-	const HEAD = 25;
-	const TAIL = 25;
-	const trim = <T,>(
-		arr: T[],
-	): { rows: T[]; gapStart: number | null; totalRows: number } => {
-		if (arr.length <= HEAD + TAIL) {
-			return { rows: arr, gapStart: null, totalRows: arr.length };
-		}
-		return {
-			rows: [...arr.slice(0, HEAD), ...arr.slice(-TAIL)],
-			gapStart: HEAD,
-			totalRows: arr.length,
-		};
-	};
-
-	if (fileType === "json") {
-		try {
-			const parsed = secureJSONParse(fileContent) as unknown;
-			const allRows = (Array.isArray(parsed) ? parsed : [parsed]) as Record<
-				string,
-				string
-			>[];
-			const headers = Object.keys(
-				(allRows[0] as Record<string, unknown>) || {},
-			);
-			const { rows, gapStart, totalRows } = trim(allRows);
-			return {
-				headers,
-				rows,
-				skippedLines: [],
-				gapStart,
-				totalRows,
-			};
-		} catch {
-			return {
-				headers: [],
-				rows: [],
-				skippedLines: [],
-				gapStart: null,
-				totalRows: 0,
-			};
-		}
-	}
-
-	const allLines = fileContent.split(/\r?\n/).filter((l) => l.trim());
-	const lines = allLines.filter((l) => {
-		const trimmed = l.trim();
-		return commentChar ? !trimmed.startsWith(commentChar) : true;
-	});
-
-	if (lines.length === 0) {
-		return {
-			headers: [],
-			rows: [],
-			skippedLines: [],
-			gapStart: null,
-			totalRows: 0,
-		};
-	}
-
-	const headerRowIndex = Math.max(0, startRow - 1);
-	const skippedLines = lines.slice(0, headerRowIndex);
-	const headerLine = lines[headerRowIndex] || "";
-	const headers = splitCSVLine(headerLine, delimiter).map((h) =>
-		h.trim().replace(/^"|"$/g, ""),
-	);
-	const allDataLines = lines.slice(headerRowIndex + 1);
-	const { rows: keptLines, gapStart, totalRows } = trim(allDataLines);
-	const rows = keptLines.map((line) =>
-		splitCSVLine(line, delimiter).map((v) => v.trim().replace(/^"|"$/g, "")),
-	);
-
-	return { headers, rows, skippedLines, gapStart, totalRows };
-}
-
-function generateColumnConfigs(
-	previewData: PreviewData,
-	columnOverrides: Record<number, Partial<ColumnConfig>>,
-	decimalPoint: string,
-	fileType: "csv" | "json" | "excel",
-): ColumnConfig[] {
-	return previewData.headers.map((name, index) => {
-		const override = columnOverrides[index];
-
-		const firstVal =
-			fileType === "json"
-				? (previewData.rows as Record<string, string>[]).find(
-						(row) => row[name],
-					)?.[name]
-				: (previewData.rows as string[][])[0]?.[index];
-
-		const { type: autoType, dateFormat: autoFormat } =
-			detectColumnTypeAndFormat(firstVal, decimalPoint);
-
-		if (override) {
-			return {
-				index,
-				name,
-				type: override.type || autoType,
-				dateFormat: override.dateFormat || autoFormat,
-				...override,
-			};
-		}
-
-		return { index, name, type: autoType, dateFormat: autoFormat };
-	});
-}
-
-function getPreferredXAxisColumn(
-	columnConfigs: ColumnConfig[],
-	xAxisColumnOverride: string | null,
-): string {
-	const nonIgnored = columnConfigs.filter((c) => c.type !== "ignore");
-	if (
-		xAxisColumnOverride &&
-		nonIgnored.some((c) => c.name === xAxisColumnOverride)
-	) {
-		return xAxisColumnOverride;
-	}
-	return (
-		nonIgnored.find((c) => c.type === "date")?.name || nonIgnored[0]?.name || ""
-	);
 }
 
 export const ImportSettingsDialog: React.FC<ImportSettingsDialogProps> = ({
